requirements.txt
discord.py
pytz
flask
ansi2html
emojis
requests
matplotlib
scipy